reorganized ingest directory
