+++ tests/PUMS/test_PUMS_economic_download.py
@@ -1,7 +1,7 @@
 """"""
 
 import pytest
-from tests.PUMS_Tests.local_loader import LocalLoader
+from tests.PUMS.local_loader import LocalLoader
 
 local_loader = LocalLoader()
