+++ tests/PUMS/test_PUMS_median_demographics_aggregation.py
@@ -1,6 +1,6 @@
 import pytest
 from tests.util import races
-from tests.PUMS_Tests.local_loader import LocalLoader
+from tests.PUMS.local_loader import LocalLoader
 
 
 local_loader = LocalLoader()
